split-view-list component and equipment uses it
