+++ src/components/equipment/equipment-display.tsx
@@ -1,14 +1,13 @@
 'use client'
 
 import { Equipment } from '@/models/equipment'
+import { SortDirection } from 'ag-grid-community'
 import { useEffect, useState } from 'react'
+import SplitViewDisplay from '../split-view-list/split-view-list'
 import EquipmentCard from './equipment-card'
-import SelectableGrid from '../selectable-grid/selectable-grid'
-import CardDisplayList from '../card-display-list/card-display-list'
 
 export default function EquipmentDisplay() {
   const [equipment, setEquipment] = useState([])
-  const [cards, setCards] = useState<Equipment[]>([])
   const [columnDefs, setColumnDefs] = useState([
     {
       field: 'level',
@@ -39,40 +38,13 @@ export default function EquipmentDisplay() {
       })
   }, [])
 
-  function handleSelectedCard(item: Equipment) {
-    setCards((cards) => {
-      let index: number = cards.indexOf(item)
-      if (index === -1) {
-        return [item, ...cards]
-      } else {
-        // this shifts the selected card to the top... maybe unnecessary?
-        // what if it autoscrolled to their position???
-        let shiftedCards = cards.slice()
-        shiftedCards.splice(index, 1)
-        shiftedCards.unshift(item)
-        return shiftedCards
-      }
-    })
-  }
-
   return (
-    <div className="h-full grid grid-cols-5 gap-x-4">
-      <div className="bg-neutral-800 p-3 rounded col-span-3 shadow-slate-200 drop-shadow-md">
-        <SelectableGrid
-          rowData={equipment}
-          onSelectedItem={handleSelectedCard}
-          columnDefs={columnDefs}
-        ></SelectableGrid>
-      </div>
-      <div className="h-full bg-neutral-800 p-3 rounded overflow-y-scroll col-span-2 shadow-slate-200 drop-shadow-md">
-        <CardDisplayList
-          children={cards.map((value) => (
-            <div key={value.name} className="pb-4">
-              <EquipmentCard value={value}></EquipmentCard>
-            </div>
-          ))}
-        ></CardDisplayList>
-      </div>
-    </div>
+    <SplitViewDisplay
+      columnDefs={columnDefs}
+      entities={equipment}
+      buildCard={(equipment: Equipment) => (
+        <EquipmentCard value={equipment}></EquipmentCard>
+      )}
+    ></SplitViewDisplay>
   )
 }
